requirements.txt
python-telegram-bot
python-dotenv
telethon==1.0.0
